requirements.txt
Django >= 4.2.1,<4.3
djangorestframework>=3.14.0,<3.15
mysqlclient>=2.1.1,<2.2
drf-spectacular>=0.26.2,<0.27
djangorestframework-simplejwt>=5.2.2,<5.3
pandas>=2.0.2,<2.1
celery>=5.3.1,<5.4
redis>=4.6.0,<4.7
django-cors-headers>=4.1.0,<4.2
openpyxl>=3.1.2,<3.2
Pillow==10.0.1
